+++ isomorph.js
@@ -1,5 +1,5 @@
 /**
- * isomorph 0.1.6 - https://github.com/insin/isomorph
+ * isomorph 0.1.7 - https://github.com/insin/isomorph
  * MIT Licensed
  */
 ;(function() {
@@ -248,6 +248,14 @@ function lookup(arr) {
   return obj
 }
 
+/**
+ * If the given object has the given property, returns its value, otherwise
+ * returns the given default value.
+ */
+function get(obj, prop, defaultValue) {
+  return (hasOwn(obj, prop) ? obj[prop] : defaultValue)
+}
+
 module.exports = {
   hasOwn: hasOwn
 , extend: extend
@@ -255,6 +263,7 @@ module.exports = {
 , items: items
 , fromItems: fromItems
 , lookup: lookup
+, get: get
 }
 })
 
@@ -350,8 +359,660 @@ module.exports = {
 }
 })
 
+require.define("./copy", function(module, exports, require) {
+var is = require('./is')
+
+/* This file is part of OWL JavaScript Utilities.
+
+OWL JavaScript Utilities is free software: you can redistribute it and/or
+modify it under the terms of the GNU Lesser General Public License
+as published by the Free Software Foundation, either version 3 of
+the License, or (at your option) any later version.
+
+OWL JavaScript Utilities is distributed in the hope that it will be useful,
+but WITHOUT ANY WARRANTY; without even the implied warranty of
+MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+GNU Lesser General Public License for more details.
+
+You should have received a copy of the GNU Lesser General Public
+License along with OWL JavaScript Utilities.  If not, see
+<http://www.gnu.org/licenses/>.
+*/
+
+// Re-usable constructor function used by clone()
+function Clone() {}
+
+// Clone objects, skip other types
+function clone(target) {
+  if (typeof target == 'object') {
+    Clone.prototype = target
+    return new Clone()
+  }
+  else {
+    return target
+  }
+}
+
+// Shallow Copy
+function copy(target) {
+  if (typeof target != 'object') {
+    // Non-objects have value semantics, so target is already a copy
+    return target
+  }
+  else {
+    var value = target.valueOf()
+    if (target != value) {
+      // the object is a standard object wrapper for a native type, say String.
+      // we can make a copy by instantiating a new object around the value.
+      return new target.constructor(value)
+    }
+    else {
+      // We have a normal object. If possible, we'll clone the original's
+      // prototype (not the original) to get an empty object with the same
+      // prototype chain as the original. If just copy the instance properties.
+      // Otherwise, we have to copy the whole thing, property-by-property.
+      if (target instanceof target.constructor && target.constructor !== Object) {
+        var c = clone(target.constructor.prototype)
+
+        // Give the copy all the instance properties of target. It has the same
+        // prototype as target, so inherited properties are already there.
+        for (var property in target) {
+          if (target.hasOwnProperty(property)) {
+            c[property] = target[property]
+          }
+        }
+      }
+      else {
+        var c = {}
+        for (var property in target) {
+          c[property] = target[property]
+        }
+      }
+
+      return c
+    }
+  }
+}
+
+// Deep Copy
+var deepCopiers = []
+
+function DeepCopier(config) {
+  for (var key in config) {
+    this[key] = config[key]
+  }
+}
+
+DeepCopier.prototype = {
+  constructor: DeepCopier
+
+  // Determines if this DeepCopier can handle the given object.
+, canCopy: function(source) { return false }
+
+  // Starts the deep copying process by creating the copy object. You can
+  // initialize any properties you want, but you can't call recursively into the
+  // DeepCopyAlgorithm.
+, create: function(source) {}
+
+  // Completes the deep copy of the source object by populating any properties
+  // that need to be recursively deep copied. You can do this by using the
+  // provided deepCopyAlgorithm instance's deepCopy() method. This will handle
+  // cyclic references for objects already deepCopied, including the source
+  // object itself. The "result" passed in is the object returned from create().
+, populate: function(deepCopyAlgorithm, source, result) {}
+}
+
+function DeepCopyAlgorithm() {
+  // copiedObjects keeps track of objects already copied by this deepCopy
+  // operation, so we can correctly handle cyclic references.
+  this.copiedObjects = []
+  var thisPass = this
+  this.recursiveDeepCopy = function(source) {
+    return thisPass.deepCopy(source)
+  }
+  this.depth = 0
+}
+DeepCopyAlgorithm.prototype = {
+  constructor: DeepCopyAlgorithm
+
+, maxDepth: 256
+
+  // Add an object to the cache.  No attempt is made to filter duplicates; we
+  // always check getCachedResult() before calling it.
+, cacheResult: function(source, result) {
+    this.copiedObjects.push([source, result])
+  }
+
+  // Returns the cached copy of a given object, or undefined if it's an object
+  // we haven't seen before.
+, getCachedResult: function(source) {
+    var copiedObjects = this.copiedObjects
+    var length = copiedObjects.length
+    for ( var i=0; i<length; i++ ) {
+      if ( copiedObjects[i][0] === source ) {
+        return copiedObjects[i][1]
+      }
+    }
+    return undefined
+  }
+
+  // deepCopy handles the simple cases itself: non-objects and object's we've
+  // seen before. For complex cases, it first identifies an appropriate
+  // DeepCopier, then calls applyDeepCopier() to delegate the details of copying
+  // the object to that DeepCopier.
+, deepCopy: function(source) {
+    // null is a special case: it's the only value of type 'object' without
+    // properties.
+    if (source === null) return null
+
+    // All non-objects use value semantics and don't need explict copying
+    if (typeof source != 'object') return source
+
+    var cachedResult = this.getCachedResult(source)
+
+    // We've already seen this object during this deep copy operation so can
+    // immediately return the result. This preserves the cyclic reference
+    // structure and protects us from infinite recursion.
+    if (cachedResult) return cachedResult
+
+    // Objects may need special handling depending on their class. There is a
+    // class of handlers call "DeepCopiers" that know how to copy certain
+    // objects. There is also a final, generic deep copier that can handle any
+    // object.
+    for (var i=0; i<deepCopiers.length; i++) {
+      var deepCopier = deepCopiers[i]
+      if (deepCopier.canCopy(source)) {
+        return this.applyDeepCopier(deepCopier, source)
+      }
+    }
+    // The generic copier can handle anything, so we should never reach this
+    // line.
+    throw new Error('no DeepCopier is able to copy ' + source)
+  }
+
+  // Once we've identified which DeepCopier to use, we need to call it in a
+  // very particular order: create, cache, populate.This is the key to detecting
+  // cycles. We also keep track of recursion depth when calling the potentially
+  // recursive populate(): this is a fail-fast to prevent an infinite loop from
+  // consuming all available memory and crashing or slowing down the browser.
+, applyDeepCopier: function(deepCopier, source) {
+    // Start by creating a stub object that represents the copy.
+    var result = deepCopier.create(source)
+
+    // We now know the deep copy of source should always be result, so if we
+    // encounter source again during this deep copy we can immediately use
+    // result instead of descending into it recursively.
+    this.cacheResult(source, result)
+
+    // Only DeepCopier.populate() can recursively deep copy.  o, to keep track
+    // of recursion depth, we increment this shared counter before calling it,
+    // and decrement it afterwards.
+    this.depth++
+    if (this.depth > this.maxDepth) {
+      throw new Error("Exceeded max recursion depth in deep copy.")
+    }
+
+    // It's now safe to let the deepCopier recursively deep copy its properties
+    deepCopier.populate(this.recursiveDeepCopy, source, result)
+
+    this.depth--
+
+    return result
+  }
+}
+
+// Entry point for deep copy.
+//   source is the object to be deep copied.
+//   maxDepth is an optional recursion limit. Defaults to 256.
+function deepCopy(source, maxDepth) {
+  var deepCopyAlgorithm = new DeepCopyAlgorithm()
+  if (maxDepth) {
+    deepCopyAlgorithm.maxDepth = maxDepth
+  }
+  return deepCopyAlgorithm.deepCopy(source)
+}
+
+// Publicly expose the DeepCopier class
+deepCopy.DeepCopier = DeepCopier
+
+// Publicly expose the list of deepCopiers
+deepCopy.deepCopiers = deepCopiers
+
+// Make deepCopy() extensible by allowing others to register their own custom
+// DeepCopiers.
+deepCopy.register = function(deepCopier) {
+  if (!(deepCopier instanceof DeepCopier)) {
+    deepCopier = new DeepCopier(deepCopier)
+  }
+  deepCopiers.unshift(deepCopier)
+}
+
+// Generic Object copier
+// The ultimate fallback DeepCopier, which tries to handle the generic case.
+// This should work for base Objects and many user-defined classes.
+deepCopy.register({
+  canCopy: function(source) { return true }
+
+, create: function(source) {
+    if (source instanceof source.constructor) {
+      return clone(source.constructor.prototype)
+    }
+    else {
+      return {}
+    }
+  }
+
+, populate: function(deepCopy, source, result) {
+    for (var key in source) {
+      if (source.hasOwnProperty(key)) {
+        result[key] = deepCopy(source[key])
+      }
+    }
+    return result
+  }
+})
+
+// Array copier
+deepCopy.register({
+  canCopy: function(source) {
+    return is.Array(source)
+  }
+
+, create: function(source) {
+    return new source.constructor()
+  }
+
+, populate: function(deepCopy, source, result) {
+    for (var i = 0; i < source.length; i++) {
+      result.push(deepCopy(source[i]))
+    }
+    return result
+  }
+})
+
+// Date copier
+deepCopy.register({
+  canCopy: function(source) {
+    return is.Date(source)
+  }
+
+, create: function(source) {
+    return new Date(source)
+  }
+})
+
+// RegExp copier
+deepCopy.register({
+  canCopy: function(source) {
+    return is.RegExp(source)
+  }
+
+, create: function(source) {
+    return source
+  }
+})
+
+module.exports = {
+  DeepCopyAlgorithm: DeepCopyAlgorithm
+, copy: copy
+, clone: clone
+, deepCopy: deepCopy
+}
+})
+
+require.define("./time", function(module, exports, require) {
+var is = require('./is')
+
+/**
+ * Pads a number with a leading zero if necessary.
+ */
+function pad(number) {
+  return (number < 10 ? '0' + number : number)
+}
+
+/**
+ * Returns the index of item in list, or -1 if it's not in list.
+ */
+function indexOf(item, list) {
+  for (var i = 0, l = list.length; i < l; i++) {
+    if (item === list[i]) {
+      return i
+    }
+  }
+  return -1
+}
+
+/**
+ * Maps directive codes to regular expression patterns which will capture the
+ * data the directive corresponds to, or in the case of locale-dependent
+ * directives, a function which takes a locale and generates a regular
+ * expression pattern.
+ */
+var parserDirectives = {
+  // Locale's abbreviated month name
+  'b': function(l) { return '(' + l.b.join('|') + ')' }
+  // Locale's full month name
+, 'B': function(l) { return '(' + l.B.join('|') + ')' }
+  // Locale's equivalent of either AM or PM.
+, 'p': function(l) { return '(' + l.AM + '|' + l.PM + ')' }
+, 'd': '(\\d\\d?)' // Day of the month as a decimal number [01,31]
+, 'H': '(\\d\\d?)' // Hour (24-hour clock) as a decimal number [00,23]
+, 'I': '(\\d\\d?)' // Hour (12-hour clock) as a decimal number [01,12]
+, 'm': '(\\d\\d?)' // Month as a decimal number [01,12]
+, 'M': '(\\d\\d?)' // Minute as a decimal number [00,59]
+, 'S': '(\\d\\d?)' // Second as a decimal number [00,59]
+, 'y': '(\\d\\d?)' // Year without century as a decimal number [00,99]
+, 'Y': '(\\d{4})'  // Year with century as a decimal number
+, '%': '%'         // A literal '%' character
+}
+
+/**
+ * Maps directive codes to functions which take the date to be formatted and
+ * locale details (if required), returning an appropriate formatted value.
+ */
+var formatterDirectives = {
+  'a': function(d, l) { return l.a[d.getDay()] }
+, 'A': function(d, l) { return l.A[d.getDay()] }
+, 'b': function(d, l) { return l.b[d.getMonth()] }
+, 'B': function(d, l) { return l.B[d.getMonth()] }
+, 'd': function(d) { return pad(d.getDate(), 2) }
+, 'H': function(d) { return pad(d.getHours(), 2) }
+, 'M': function(d) { return pad(d.getMinutes(), 2) }
+, 'm': function(d) { return pad(d.getMonth() + 1, 2) }
+, 'S': function(d) { return pad(d.getSeconds(), 2) }
+, 'w': function(d) { return d.getDay() }
+, 'Y': function(d) { return d.getFullYear() }
+, '%': function(d) { return '%' }
+}
+
+/** Test for hanging percentage symbols. */
+var strftimeFormatCheck = /[^%]%$/
+
+/**
+ * A partial implementation of strptime which parses time details from a string,
+ * based on a format string.
+ * @param {String} format
+ * @param {Object} locale
+ */
+function TimeParser(format, locale) {
+  this.format = format
+  this.locale = locale
+  var cachedPattern = TimeParser._cache[locale.name + '|' + format]
+  if (cachedPattern !== undefined) {
+    this.re = cachedPattern[0]
+    this.matchOrder = cachedPattern[1]
+  }
+  else {
+    this.compilePattern()
+  }
+}
+
+/**
+ * Caches RegExps and match orders generated per locale/format string combo.
+ */
+TimeParser._cache = {}
+
+TimeParser.prototype.compilePattern = function() {
+  // Normalise whitespace before further processing
+  var format = this.format.split(/(?:\s|\t|\n)+/).join(' ')
+    , pattern = []
+    , matchOrder = []
+    , c
+    , directive
+
+  for (var i = 0, l = format.length; i < l; i++) {
+    c = format.charAt(i)
+    if (c != '%') {
+      if (c === ' ') {
+        pattern.push(' +')
+      }
+      else {
+        pattern.push(c)
+      }
+      continue
+    }
+
+    if (i == l - 1) {
+      throw new Error('strptime format ends with raw %')
+    }
+
+    c = format.charAt(++i)
+    directive = parserDirectives[c]
+    if (directive === undefined) {
+      throw new Error('strptime format contains an unknown directive: %' + c)
+    }
+    else if (is.Function(directive)) {
+      pattern.push(directive(this.locale))
+    }
+    else {
+      pattern.push(directive)
+    }
+
+    if (c != '%') {
+       matchOrder.push(c)
+    }
+  }
+
+  this.re = new RegExp('^' + pattern.join('') + '$')
+  this.matchOrder = matchOrder
+  TimeParser._cache[this.locale.name + '|' + this.format] = [this.re, matchOrder]
+}
+
+/**
+ * Attempts to extract date and time details from the given input.
+ * @param {string} input
+ * @return {Array.<number>}
+ */
+TimeParser.prototype.parse = function(input) {
+  var matches = this.re.exec(input)
+  if (matches === null) {
+    throw new Error('Time data did not match format: data=' + input +
+                    ', format=' + this.format)
+  }
+
+    // Default values for when more accurate values cannot be inferred
+  var time = [1900, 1, 1, 0, 0, 0]
+    // Matched time data, keyed by directive code
+    , data = {}
+
+  for (var i = 1, l = matches.length; i < l; i++) {
+    data[this.matchOrder[i - 1]] = matches[i]
+  }
+
+  // Extract year
+  if (data.hasOwnProperty('Y')) {
+    time[0] = parseInt(data.Y, 10)
+  }
+  else if (data.hasOwnProperty('y')) {
+    var year = parseInt(data.y, 10)
+    if (year < 68) {
+        year = 2000 + year
+    }
+    else if (year < 100) {
+        year = 1900 + year
+    }
+    time[0] = year
+  }
+
+  // Extract month
+  if (data.hasOwnProperty('m')) {
+    var month = parseInt(data.m, 10)
+    if (month < 1 || month > 12) {
+      throw new Error('Month is out of range: ' + month)
+    }
+    time[1] = month
+  }
+  else if (data.hasOwnProperty('B')) {
+    time[1] = indexOf(data.B, this.locale.B) + 1
+  }
+  else if (data.hasOwnProperty('b')) {
+    time[1] = indexOf(data.b, this.locale.b) + 1
+  }
+
+  // Extract day of month
+  if (data.hasOwnProperty('d')) {
+    var day = parseInt(data.d, 10)
+    if (day < 1 || day > 31) {
+      throw new Error('Day is out of range: ' + day)
+    }
+    time[2] = day
+  }
+
+  // Extract hour
+  if (data.hasOwnProperty('H')) {
+    var hour = parseInt(data.H, 10)
+    if (hour > 23) {
+      throw new Error('Hour is out of range: ' + hour)
+    }
+    time[3] = hour
+  }
+  else if (data.hasOwnProperty('I')) {
+    var hour = parseInt(data.I, 10)
+    if (hour < 1 || hour > 12) {
+      throw new Error('Hour is out of range: ' + hour)
+    }
+
+    // If we don't get any more information, we'll assume this time is
+    // a.m. - 12 a.m. is midnight.
+    if (hour == 12) {
+        hour = 0
+    }
+
+    time[3] = hour
+
+    if (data.hasOwnProperty('p')) {
+      if (data.p == this.locale.PM) {
+        // We've already handled the midnight special case, so it's
+        // safe to bump the time by 12 hours without further checks.
+        time[3] = time[3] + 12
+      }
+    }
+  }
+
+  // Extract minute
+  if (data.hasOwnProperty('M')) {
+    var minute = parseInt(data.M, 10)
+    if (minute > 59) {
+        throw new Error('Minute is out of range: ' + minute)
+    }
+    time[4] = minute
+  }
+
+  // Extract seconds
+  if (data.hasOwnProperty('S')) {
+    var second = parseInt(data.S, 10)
+    if (second > 59) {
+      throw new Error('Second is out of range: ' + second)
+    }
+    time[5] = second
+  }
+
+  // Validate day of month
+  var day = time[2], month = time[1], year = time[0]
+  if (((month == 4 || month == 6 || month == 9 || month == 11) &&
+      day > 30) ||
+      (month == 2 && day > ((year % 4 == 0 && year % 100 != 0 ||
+                             year % 400 == 0) ? 29 : 28))) {
+    throw new Error('Day is out of range: ' + day)
+  }
+
+  return time
+}
+
+var time  = {
+  /** Default locale name. */
+  defaultLocale: 'en'
+
+  /** Locale details. */
+, locales: {
+    en: {
+      name: 'en'
+    , a: ['Sun', 'Mon', 'Tue', 'Wed', 'Thu', 'Fri', 'Sat']
+    , A: ['Sunday', 'Monday', 'Tuesday', 'Wednesday', 'Thursday',
+          'Friday', 'Saturday']
+    , AM: 'AM'
+    , b: ['Jan', 'Feb', 'Mar', 'Apr', 'May', 'Jun', 'Jul', 'Aug', 'Sep',
+          'Oct', 'Nov', 'Dec']
+    , B: ['January', 'February', 'March', 'April', 'May', 'June', 'July',
+          'August', 'September', 'October', 'November', 'December']
+    , PM: 'PM'
+    }
+  }
+}
+
+/**
+ * Retrieves the locale with the given code.
+ * @param {string} code
+ * @return {Object}
+ */
+var getLocale = time.getLocale = function(code) {
+  if (code) {
+    if (time.locales.hasOwnProperty(code)) {
+      return time.locales[code]
+    }
+    else if (code.length > 2) {
+      // If we appear to have more than a language code, try the
+      // language code on its own.
+      var languageCode = code.substring(0, 2)
+      if (time.locales.hasOwnProperty(languageCode)) {
+        return time.locales[languageCode]
+      }
+    }
+  }
+  return time.locales[time.defaultLocale]
+}
+
+/**
+ * Parses time details from a string, based on a format string.
+ * @param {string} input
+ * @param {string} format
+ * @param {string=} locale
+ * @return {Array.<number>}
+ */
+var strptime = time.strptime = function(input, format, locale) {
+  return new TimeParser(format, getLocale(locale)).parse(input)
+}
+
+/**
+ * Convenience wrapper around time.strptime which returns a JavaScript Date.
+ * @param {string} input
+ * @param {string} format
+ * @param {string=} locale
+ * @return {date}
+ */
+time.strpdate = function(input, format, locale) {
+  var t = strptime(input, format, locale)
+  return new Date(t[0], t[1] - 1, t[2], t[3], t[4], t[5])
+}
+
+/**
+ * A partial implementation of <code>strftime</code>, which formats a date
+ * according to a format string. An Error will be thrown if an invalid
+ * format string is given.
+ * @param {date} date
+ * @param {string} format
+ * @param {string=} locale
+ * @return {string}
+ */
+time.strftime = function(date, format, locale) {
+  if (strftimeFormatCheck.test(format)) {
+    throw new Error('strftime format ends with raw %')
+  }
+  locale = getLocale(locale)
+  return format.replace(/(%.)/g, function(s, f) {
+    var code = f.charAt(1)
+    if (typeof formatterDirectives[code] == 'undefined') {
+      throw new Error('strftime format contains an unknown directive: ' + f)
+    }
+    return formatterDirectives[code](date, locale)
+  })
+}
+
+module.exports = time
+})
+
 require.define("isomorph", function(module, exports, require) {
-exports.version = '0.1.5'
+exports.version = '0.1.7'
 
 exports.is = require('./is')
 exports.array = require('./array')
@@ -360,6 +1021,8 @@ exports.object = require('./object')
 exports.format = require('./format')
 exports.re = require('./re')
 exports.querystring = require('./querystring')
+exports.copy = require('./copy')
+exports.time = require('./time')
 })
 
 window['isomorph'] = require('isomorph')