+++ CHANGES.rst
@@ -1,3 +1,11 @@
+0.1.7 / 2012-01-28
+==================
+
+* Added time module, with``time.strftime``, ``time.strpdate``, ``time.strftime``
+  and some basic locale support.
+* Added copy module, with ``copy.copy()`` and ``copy.deepCopy()``
+* Added ``object.get()``
+
 0.1.6 / 2012-01-25
 ==================